requirements.txt
telethon>=1.28.5
Pillow>=9.5.0
matplotlib>=3.7.1
numpy>=1.24.3
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.15 
aportalsmp

# Local package
-e .

aiogram>=3.0.0
python-dotenv>=0.19.0
Pillow>=8.3.1
requests>=2.26.0
pyrogram>=2.0.0
aportalsmp